[typescript]: remove fuzz blockers and cleanup harnesses
fix: add header
