+++ projects/typescript/Dockerfile
@@ -19,10 +19,12 @@ FROM gcr.io/oss-fuzz-base/base-builder-javascript
 RUN  git clone --depth 1 https://github.com/microsoft/TypeScript.git
 
 COPY build.sh $SRC/
+COPY fuzz_util.js $SRC/TypeScript
 COPY fuzz_ast.js $SRC/TypeScript
 COPY fuzz_compiler.js $SRC/TypeScript
 COPY fuzz_scanner.js $SRC/TypeScript
 COPY fuzz_json_parser.js $SRC/TypeScript
 COPY fuzz_transpile_module.js $SRC/TypeScript
+
 
 WORKDIR $SRC/TypeScript
